Paris Neighbourhood Geo-Coordinates
Data for collaboration and to expanding the business problem of relocation options used here. I hope you will fork the Github gist and/or repository and collaborate!!! It was prepared for an assignment for a course on Coursera called Applied Data Science Capstone, you can take this course online by clicking here.