Relocation Options
 Cousera (Applied Data Science) Capstone Project

This IPython notebook contains code and is annotated to answer question
from the first part of the above assignment. Here we focus on how to search
for, select, mine and upload potential data as well as store and clean
dataframes with pandas and nympy. The outcome and notebook itself, will be
used in the creation of a deliverable, a blog post. The blog post and other
deliverables are part of the later half of the assignment and will be
available a week after this submission.